delete_course.php: csrf

+++ modules/course_info/delete_course.php
@@ -33,6 +33,18 @@ $tool_content = "";
 
 if ($is_adminOfCourse) {
     if (isset($delete)) {
+        // csrf
+        if (!isset($_SESSION['token']) || !isset($_GET['token'])) {
+            header("location:" . $_SERVER['PHP_SELF']);
+            exit();
+        }
+
+        if ($_SESSION['token'] !== $_GET['token']) {
+            header("location:" . $_SERVER['PHP_SELF']);
+            exit();
+        }
+        unset($_SESSION['token']);
+
         mysql_select_db("$mysqlMainDb", $db);
         mysql_query("DROP DATABASE `$currentCourseID`");
         mysql_query("DELETE FROM `$mysqlMainDb`.cours WHERE code='$currentCourseID'");
@@ -51,6 +63,7 @@ if ($is_adminOfCourse) {
         draw($tool_content, 1);
         exit();
     } else {
+        $form_token = $_SESSION['token'] = md5(mt_rand());
         $tool_content .= "
 		<table width=\"99%\">
 		<tbody>
@@ -61,7 +74,7 @@ if ($is_adminOfCourse) {
 		</tr>
 		<tr>
 		<th rowspan='2' class='left' width='220'>$langConfirmDel :</th>
-		<td width='52' align='center'><a href=\"" . $_SERVER['PHP_SELF'] . "?delete=yes\">$langYes</a></td>
+		<td width='52' align='center'><a href=\"" . $_SERVER['PHP_SELF'] . "?delete=yes&token=$form_token\">$langYes</a></td>
 		<td><small>$langByDel</small></td>
 		</tr>
 		<tr>
